chore: move all election data files to election folder
