+++ src/data/election/2012.ts
