+++ src/data/election/2020.ts
