+++ src/data/election/2024.ts
